pyproject.toml
[project]
name = "renfeserver"
version = "0.1.0"
description = "Renfe Server Backend"
authors = [{name = "Juan Macias", email = "extjmv@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "alembic>=1.13.2",
    "bcrypt>=4.3.0",
    "dependency-injector>=4.48.1",
    "fastapi>=0.115.0",
    "python-multipart>=0.0.20",
    "email-validator>=2.0.0",
    "httptools>=0.6.4",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.3.4",
    "python-jose>=3.3.0",
    "python-dotenv>=1.0.0",
    "ulid-py>=1.1.0",
    "uvloop>=0.21.0",
    "watchfiles>=1.1.0",
    "websockets>=15.0.1",
    "uvicorn>=0.35.0",
    # Reporting dependencies
    "celery>=5.3.0",
    "redis>=5.0.0",
    "weasyprint>=60.0",
    "boto3>=1.34.0",
    "jinja2>=3.1.0",
    "pyjwt>=2.10.1",
    "httpx>=0.28.1",
    "sentry-sdk[fastapi]>=2.0.0",
    "stripe>=7.0.0",
    "beautifulsoup4>=4.14.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.1",
    "faker>=30.8.2",
    "mypy>=1.17.1",
    "flake8>=7.3.0",
    "autopep8>=2.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
